feat: esbuild runtime (#122)
Using arc runtimes we're able to have a cleaner build pipeline. It looks like hydration and other features will come to runtimes one day too.

This update also bring esbuild plugin support via the configFile directive.

BREAKING CHANGE: esbuild is now a runtime and the runtime must be set for this to work.